2004-03-07  Daniel Berlin  <dberlin@dberlin.org>
	* engine/Makefile.am (AM_CFLAGS): Add -I$(top_srcddir)/../include, so we
	can get ansidecl.h and friends.
	* engine/array.c (array_extend): Fix unsigned/signed comparison.
	* engine/banshee.h (struct gen_e): Just make void * when empty, instead
	of empty struct.
	* engine/compiler.h: Don't ever define HAVE_VARIADIC_MACROS, -pedantic
	whines about them even when we know we can use them.
	* engine/dot.c: constify our char *'s everywhere.
	(declare_node): Cast &result to char * before casting to hash_data *.
	* engine/dot.h: constify our char *'s everywhere.
	* engine/flow-var.c: Ditto.
	* engine/flow-var.h: Ditto.
	* engine/flowrow-sort.c: Include ansidecl.h.
	(get_contour): Mark unused parameter.
	(update_upper_bound): Make comment /* instead of //.
	(field_eq): Staticify.
	* engine/hash.c (string_hash): Staticify.
	(string_eq): Ditto.
	(entry_cmp): Ditto.
	* engine/hash.h: Move comments around.
	(hash_table_apply): Add prototype.
	* engine/hashset.c (INIT_TABLE_SIZE): Remove extra semicolon.
	(EMPTY_KEY): Ditto.
	* engine/jcollection.c (jcoll_create_chain): // -> /* comment.
	(jcoll_accum): Staticify.
	* engine/list.c (sort_linked_list): Staticify.
	(compare): Use comparator_fn as last argument, not void *.
	* engine/list.h: Declare the functions even for opaque lists.
	(list_clear macro): don't return value.
	* engine/setif-sort.c: Replaced with new version from banshee cvs.
	* engine/setif-var.c: constify our char *'s.
	* engine/setif-var.h: Ditto.
	* engine/setst-var.c: Ditto.
	* engine/setst-var.h: Ditto.
	* engine/stamp.c (stamp_string): Use long, not int.
	Cast through char *.
	* engine/stamp.h (stamp): Long, not int.
	* engine/term-sort.c (term_print_stats): Mark argument unused.
	include ansidecl.h.
	* engine/term-var.c: constify the char *'s.
	* engine/term-var.h: Ditto.
	* engine/ufind.h (union find update macro): Don't return uf_update's value
	(which is void anyway).
	* engine/util.c (ptr_hash): Return long, not int.
	* engine/util.h (ptr_hash): Ditto.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/tree-ssa-20020619-branch@79078 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libbanshee/engine/util.h
@@ -87,7 +87,7 @@ char *rvsprintf(region r, const char *fmt, va_list args);
 char *ptr_to_ascii(void *ptr);
 
 /* Convert a pointer to an integer */
-int ptr_hash(void *ptr);
+long ptr_hash(void *ptr);
 
 /* Return TRUE iff ptr1 == ptr2 */
 bool ptr_eq(void *ptr1, void *ptr2);
